Create new route and update links
